refactor(magic): split services.py into thematic submodules
Convert world/magic/services.py into world/magic/services/ package
per Scope 6 §4.4. __init__.py re-exports full public surface.
