+++ src/world/magic/services/anima.py
@@ -1,0 +1,30 @@
+"""Anima resource service functions for the magic system."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from django.db import transaction
+
+from world.magic.models import CharacterAnima
+
+if TYPE_CHECKING:
+    from evennia.objects.models import ObjectDB
+
+
+def deduct_anima(character: ObjectDB, effective_cost: int) -> int:
+    """Deduct anima from character, returning the overburn deficit.
+
+    Uses select_for_update inside transaction.atomic for race-condition
+    safety, following the ActionPointPool.spend() pattern.
+    Returns 0 if no overburn, positive int if life force is drawn.
+    """
+    if effective_cost <= 0:
+        return 0
+
+    with transaction.atomic():
+        anima = CharacterAnima.objects.select_for_update().get(character=character)
+        deficit = max(effective_cost - anima.current, 0)
+        anima.current = max(anima.current - effective_cost, 0)
+        anima.save(update_fields=["current"])
+    return deficit
