+++ src/world/magic/services/soulfray.py
@@ -1,0 +1,241 @@
+"""Soulfray accumulation, severity, warning, and mishap service functions."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from world.magic.models import SoulfrayConfig
+from world.magic.types import MishapResult, SoulfrayResult, SoulfrayWarning
+
+if TYPE_CHECKING:
+    from evennia.objects.models import ObjectDB
+
+    from actions.models.action_templates import ConsequencePool
+    from world.checks.types import CheckResult
+
+
+def calculate_soulfray_severity(
+    current_anima: int,
+    max_anima: int,
+    deficit: int,
+    config: SoulfrayConfig,
+) -> int:
+    """Compute Soulfray severity contribution from post-deduction anima state."""
+    from decimal import Decimal  # noqa: PLC0415
+    from math import ceil  # noqa: PLC0415
+
+    if max_anima <= 0:
+        return 0
+
+    ratio = Decimal(current_anima) / Decimal(max_anima)
+    threshold = config.soulfray_threshold_ratio
+
+    if ratio >= threshold:
+        return 0
+
+    depletion = float((threshold - ratio) / threshold)
+    severity = ceil(config.severity_scale * depletion)
+
+    if deficit > 0:
+        severity += ceil(config.deficit_scale * deficit)
+
+    return severity
+
+
+def get_soulfray_warning(character: ObjectDB) -> SoulfrayWarning | None:
+    """Return the current Soulfray stage warning for the safety checkpoint."""
+    from world.conditions.models import ConditionInstance  # noqa: PLC0415
+    from world.magic.audere import SOULFRAY_CONDITION_NAME  # noqa: PLC0415
+
+    soulfray_instance = (
+        ConditionInstance.objects.filter(
+            target=character,
+            condition__name=SOULFRAY_CONDITION_NAME,
+        )
+        .select_related("current_stage", "current_stage__consequence_pool")
+        .first()
+    )
+
+    if soulfray_instance is None or soulfray_instance.current_stage is None:
+        return None
+
+    stage = soulfray_instance.current_stage
+    has_death_risk = False
+    if stage.consequence_pool_id:
+        from world.checks.models import Consequence  # noqa: PLC0415
+
+        has_death_risk = Consequence.objects.filter(
+            pool_entries__pool=stage.consequence_pool,
+            character_loss=True,
+        ).exists()
+
+    return SoulfrayWarning(
+        stage_name=stage.name,
+        stage_description=stage.description,
+        has_death_risk=has_death_risk,
+    )
+
+
+def select_mishap_pool(control_deficit: int) -> ConsequencePool | None:
+    """Select a control mishap consequence pool based on deficit magnitude."""
+    from django.db import models as db_models  # noqa: PLC0415
+
+    from world.magic.models import MishapPoolTier  # noqa: PLC0415
+
+    tier = (
+        MishapPoolTier.objects.filter(min_deficit__lte=control_deficit)
+        .filter(
+            db_models.Q(max_deficit__gte=control_deficit) | db_models.Q(max_deficit__isnull=True),
+        )
+        .order_by("-min_deficit")
+        .first()
+    )
+    return tier.consequence_pool if tier else None
+
+
+def _handle_soulfray_accumulation(
+    *,
+    character: ObjectDB,
+    soulfray_severity: int,
+    soulfray_config: SoulfrayConfig,
+    technique_check_result: CheckResult | None,
+) -> SoulfrayResult:
+    """Handle Soulfray severity accumulation, stage advancement, and consequence pool."""
+    from world.checks.consequence_resolution import (  # noqa: PLC0415
+        apply_resolution,
+        select_consequence_from_result,
+    )
+    from world.checks.services import perform_check  # noqa: PLC0415
+    from world.checks.types import ResolutionContext  # noqa: PLC0415
+    from world.conditions.models import (  # noqa: PLC0415
+        ConditionCheckModifier,
+        ConditionInstance,
+        ConditionTemplate,
+    )
+    from world.conditions.services import (  # noqa: PLC0415
+        advance_condition_severity,
+        apply_condition,
+    )
+    from world.magic.audere import SOULFRAY_CONDITION_NAME  # noqa: PLC0415
+    from world.magic.models import TechniqueOutcomeModifier  # noqa: PLC0415
+
+    # Find or create Soulfray condition
+    soulfray_instance = (
+        ConditionInstance.objects.filter(
+            target=character,
+            condition__name=SOULFRAY_CONDITION_NAME,
+        )
+        .select_related("current_stage")
+        .first()
+    )
+
+    if soulfray_instance is None:
+        soulfray_template = ConditionTemplate.objects.get(
+            name=SOULFRAY_CONDITION_NAME,
+        )
+        result = apply_condition(target=character, condition=soulfray_template)
+        soulfray_instance = result.instance
+        # apply_condition creates with severity=1. Use advance_condition_severity
+        # to set the real severity and resolve the correct stage.
+        advance_condition_severity(soulfray_instance, soulfray_severity - 1)
+        soulfray_instance.refresh_from_db()
+
+        # On first creation the pool is not fired; callers must trigger a second
+        # accumulation for the stage threshold to evaluate.
+        return SoulfrayResult(
+            severity_added=soulfray_severity,
+            stage_name=(
+                soulfray_instance.current_stage.name if soulfray_instance.current_stage else None
+            ),
+            stage_advanced=soulfray_instance.current_stage is not None,
+        )
+
+    # Advance existing condition
+    advance_result = advance_condition_severity(soulfray_instance, soulfray_severity)
+    soulfray_instance.refresh_from_db()
+
+    # Fire stage consequence pool if present
+    resilience_check = None
+    stage_consequence = None
+    current_stage = soulfray_instance.current_stage
+
+    if current_stage and current_stage.consequence_pool_id:
+        from actions.services import get_effective_consequences  # noqa: PLC0415
+
+        consequences = get_effective_consequences(
+            current_stage.consequence_pool,
+        )
+        if consequences:
+            # 1. Stage penalty via ConditionCheckModifier
+            stage_modifier = 0
+            stage_check_mod = ConditionCheckModifier.objects.filter(
+                stage=current_stage,
+                check_type=soulfray_config.resilience_check_type,
+            ).first()
+            if stage_check_mod:
+                stage_modifier = stage_check_mod.modifier_value
+
+            # 2. Technique outcome modifier (botch = penalty, crit = bonus)
+            outcome_modifier = 0
+            if technique_check_result and technique_check_result.outcome:
+                outcome_mod = TechniqueOutcomeModifier.objects.filter(
+                    outcome=technique_check_result.outcome,
+                ).first()
+                if outcome_mod:
+                    outcome_modifier = outcome_mod.modifier_value
+
+            total_modifier = stage_modifier + outcome_modifier
+
+            # Perform resilience check
+            resilience_check = perform_check(
+                character=character,
+                check_type=soulfray_config.resilience_check_type,
+                target_difficulty=soulfray_config.base_check_difficulty,
+                extra_modifiers=total_modifier,
+            )
+
+            # Select and apply consequence
+            pending = select_consequence_from_result(
+                character,
+                resilience_check,
+                consequences,
+            )
+            context = ResolutionContext(character=character)
+            applied = apply_resolution(pending, context)
+            if applied:
+                stage_consequence = applied[0]
+
+    return SoulfrayResult(
+        severity_added=soulfray_severity,
+        stage_name=current_stage.name if current_stage else None,
+        stage_advanced=advance_result.stage_changed,
+        resilience_check=resilience_check,
+        stage_consequence=stage_consequence,
+    )
+
+
+def _resolve_mishap(
+    character: ObjectDB,
+    pool: ConsequencePool,
+    check_result: CheckResult,
+) -> MishapResult | None:
+    """Resolve a mishap rider using the main check result."""
+    from actions.services import get_effective_consequences  # noqa: PLC0415
+    from world.checks.consequence_resolution import (  # noqa: PLC0415
+        apply_resolution,
+        select_consequence_from_result,
+    )
+    from world.checks.types import ResolutionContext  # noqa: PLC0415
+
+    consequences = get_effective_consequences(pool)
+    if not consequences:
+        return None
+
+    pending = select_consequence_from_result(character, check_result, consequences)
+    context = ResolutionContext(character=character)
+    applied = apply_resolution(pending, context)
+
+    return MishapResult(
+        consequence_label=pending.selected_consequence.label,
+        applied_effect_ids=[e.created_instance.pk for e in applied if e.created_instance],
+    )
